requirements.txt
requests
Pillow
python-dotenv
pytz
apscheduler
